# src/openapi/api-contract.yaml
openapi: 3.0.3
info:
  title: Education Platform API
  version: '1.0.0'
  description: >-
    Contract for the education platform API covering authentication, user
    management, course catalog and assignment submissions.
servers:
  - url: https://api.example.com/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
      description: Replay protection key. Repeat calls with the same key receive the same response.
paths:
...
  /users:
    get:
      summary: List users with pagination
      tags: [Users]
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin, moderator]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Users page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
  /users/{userId}:
    get:
      summary: Retrieve a user profile
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
    patch:
      summary: Update a user profile
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile]
              properties:
                profile:
                  $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [displayName, locale]
      properties:
        displayName:
          type: string
        locale:
          type: string
        bio:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        locale:
          type: string
        bio:
          type: string
    PublicUser:
      type: object
      required: [id, email, roles, profile]
      properties:
        id:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        profile:
          $ref: '#/components/schemas/UserProfile'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PublicUser'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
...